requirements.txt
transformers==4.51.3
peft==0.15.1
trl==0.17.0
bitsandbytes==0.45.5
accelerate==1.6.0
datasets==3.5.0
torch==2.6.0
tokenizers==0.21.1
huggingface-hub==0.30.2
wandb==0.19.10
evaluate==0.4.3
rouge-score==0.1.2
nltk==3.9.1
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.5
streamlit==1.45.0
plotly==6.0.1
python-dotenv==1.1.0
sentencepiece==0.2.0
